计时器功能

+++ FTPWinform/MainForm.cs
@@ -14,26 +14,67 @@ namespace FTPWinform
         public MainForm()
         {
             InitializeComponent();
-        }
-
-        private void button1_Click(object sender, EventArgs e)
-        {
-            FTPController.FTP ftp = new FTPController.FTP(new Uri(FTPController.Auth.Url+"/DBBackUp"), FTPController.Auth.User, FTPController.Auth.Password);
-          //  ftp.DownloadFile("memcached.rar", @"D:\");
-            FTPController.FileStruct[] files = ftp.ListFiles();
-            foreach (var file in files)
+            FTPController.MessageConsole._onLogger = (obj, level) =>
             {
-                if (ftp.DownloadFile(file.Name, @"D:\test")) {
-                    ftp.DeleteFile(file.Name);
+                if (obj != null)
+                {
+                    this.OutputLog(obj.ToString(), level);
                 }
-            }
-            FTPController.MessageConsole.WriteConsole("下载文件 完成。");
+            };
+        }
+        FTPController.Timing _timing = null;
+        private bool isRunning = false;
+        private void btnStart_Click(object sender, EventArgs e)
+        {
+            this._timing.Start();
+
         }
 
-        private void button2_Click(object sender, EventArgs e)
+        private void btnStop_Click(object sender, EventArgs e)
         {
-            //System.Net.Ftp.clsFTP ftp = new System.Net.Ftp.clsFTP(new Uri("ftp://118.186.246.162:2121"), "ftp_admin", "iwehave2305!");
-            //ftp.DeleteFile("GAIA_WEB_RETAIL_CJ_backup_2016_02_18_000001_1453809.bak");
+            this._timing.Stop();
+
+        }
+
+        private void MainForm_Load(object sender, EventArgs e)
+        {
+            DateTime startTime = DateTime.Now;
+            startTime.AddMinutes(1);
+            this._timing = new FTPController.Timing(1000, startTime);
+            this._timing._onTimingStart = () =>
+            {
+                if (!this.isRunning)
+                {
+                    this.isRunning = true;
+                    FTPController.FTP ftp = new FTPController.FTP(new Uri(FTPController.Auth.Url + "/DBBackUp"), FTPController.Auth.User, FTPController.Auth.Password);
+                    //  ftp.DownloadFile("memcached.rar", @"D:\");
+                    FTPController.FileStruct[] files = ftp.ListFiles();
+                    foreach (var file in files)
+                    {
+                        if (ftp.DownloadFile(file.Name, @"D:\test"))
+                        {
+                            ftp.DeleteFile(file.Name);
+                        }
+                    }
+                    FTPController.MessageConsole.WriteConsole("下载文件 完成。");
+                    this.isRunning = false;
+                }
+
+            };
+        }
+        public void OutputLog(string msg, FTPModel.LogModel.LogLevel level)
+        {
+            if (this.IsHandleCreated)
+            {
+                this.tbMsg.Invoke((MethodInvoker)delegate()
+                {
+                    lock (this.tbMsg)
+                    {
+                        DateTime dtime = DateTime.Now;
+                        this.tbMsg.AppendText("[" + dtime + "] " + msg + "\r\n");
+                    }
+                });
+            }
         }
     }
 }
